7. AbuseIPDB
This AbuseIPDB tool tells about how many times the IP reposted which we can get from above resources and how much abuse it is! along with all information about locations and domain names. It gives us the malicious report of an IP address. In Email Analyses IP address of sender can be checked if malicious.